separate map into different components
